pyproject.toml
[build-system]
requires = [
    "setuptools>=68.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "pet-eval"
version = "0.1.0"
requires-python = ">=3.11,<3.12"
dependencies = [
    "pet-schema @ git+https://github.com/Train-Pet-Pipeline/pet-schema.git@v1.0.0",
    "pet-infra @ git+https://github.com/Train-Pet-Pipeline/pet-infra.git@v1.0.0",
    "torch>=2.1",
    "transformers>=4.37",
    "torchaudio>=2.1",
    "peft>=0.10",
    "qwen-vl-utils",
    "bert-score>=0.3.13",
    "scipy>=1.11",
    "pyyaml",
    "tenacity",
    "python-json-logger",
]

[project.optional-dependencies]
constrained = [
    "outlines>=0.1",
]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
]

[tool.setuptools.packages.find]
where = [
    "src",
]
include = [
    "pet_eval",
    "pet_eval.*",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "N",
    "W",
    "UP",
]

[tool.mypy]
disable_error_code = ["import-untyped"]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
